wake writes after create_data_channel
